Last minutes fixes

+++ src/Exporter.h
@@ -28,11 +28,11 @@ std::string get_python_interpolation(const GradientKind& kind);
 
 std::string get_python_color(const Color& color);
 
-int sample_to_ms(int sample, int sample_rate);
+int64_t sample_to_ms(int64_t sample, int64_t sample_rate);
 
 std::string get_group_str(size_t group_id);
 
-std::string get_python_command(const Command& command, int sample_rate);
+std::string get_python_command(const Command& command, int64_t sample_rate);
 
 // Important : the keyframe list must be sorted, and every command must be retimed before generating the python script
 std::string generate_python_script(const ProjectData& data);
